# openapi.yaml
openapi: 3.1.0
info:
  title: PHC Pulse API
  version: 1.0.0
  description: |
    # PHC Pulse API Specifications
    Federated Early-Warning and Supply Chain Redistribution Platform for India's Primary Health Centres (PHCs).
    
    ## System Architecture & Non-Negotiable Rules:
    1. **Strict Service-LLM Separation**: Numerical calculations (demand forecasting, days of cover, min-cost optimization, federated metrics, impact simulations) are strictly computed by algorithmic Python microservices. Gemini models are exclusively used for reading/parsing register photos/voice notes, Q&A synthesis, transfer rationale generation, and drafting alerts.
    2. **Role & Geographic Scoping**: Data access is scoped strictly before reaching Gemini or client responses based on user role (`PHC_STAFF`, `DISTRICT_OFFICER`, `STATE_OFFICER`). Patient-level records are never captured or transmitted.
    3. **Data Sovereignty**: Physical drug redistribution is confined within state borders. Cross-state collaboration occurs exclusively through privacy-preserving Federated Model gradient updates.
    4. **Idempotency**: All state-modifying operations (such as `POST /report`) require an `idempotency_key` header or payload attribute to guarantee safe retries.
    5. **Environment Configuration**: Gemini model aliases are configured via environment variables (`GEMINI_MODEL_PARSER`, `GEMINI_MODEL_QA`, `GEMINI_MODEL_EXPLAIN`, `GEMINI_MODEL_ALERT`).
...
servers:
  - url: /api
...
paths:
  /snapshot:
    get:
      summary: Get Live Inventory Snapshot
      description: Returns real-time stock levels, demand projections, and risk levels across PHCs filtered by role scope.
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [Madhya Pradesh, Maharashtra, Kerala, Assam]
        - name: district
          in: query
          required: false
          schema:
            type: string
        - name: medicine
          in: query
          required: false
          schema:
            type: string
        - name: risk
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RiskLevel'
      responses:
        '200':
          description: Snapshot records retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SnapshotRecord'
        '400':
          $ref: '#/components/responses/400Error'
  /scenario:
    post:
      summary: Run Outbreak Demand Projection Scenario
      description: Simulates an outbreak disease surge (e.g. +40% fever surge), recomputing days of cover and risk levels across target geography.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutbreakScenarioRequest'
      responses:
        '200':
          description: Outbreak scenario executed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutbreakScenarioResult'
...
components:
  schemas:
    RiskLevel:
      type: string
      enum: [Low, Medium, High, Critical]
      description: |
        Risk level threshold based on days_of_cover:
        - Critical: <= 3.0 days
        - High: <= 7.0 days
        - Medium: <= 14.0 days
        - Low: > 14.0 days
    DiseaseClass:
      type: string
      enum: [fever_vector, diarrhoeal, respiratory, chronic_metabolic, emergency_trauma]
    SnapshotRecord:
      type: object
      required:
        - phc_id
        - phc_name
        - state
        - district
        - medicine
        - disease_class
        - stock
        - unit
        - daily_demand
        - days_of_cover
        - risk_level
        - nearest_expiry
        - cold_chain
        - beds_available
        - staff_present
        - staff_sanctioned
      properties:
        phc_id:
          type: string
          example: PHC_MP_DEW_001
        phc_name:
          type: string
          example: Rampur PHC
        state:
          type: string
          example: Madhya Pradesh
        district:
          type: string
          example: Dewas
        medicine:
          type: string
          example: Paracetamol 500mg tab
        disease_class:
          $ref: '#/components/schemas/DiseaseClass'
        stock:
          type: number
          example: 120
        unit:
          type: string
          example: tablets
        daily_demand:
          type: number
          example: 30
        days_of_cover:
          type: number
          example: 4.0
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
        nearest_expiry:
          type: string
          example: "2026-11-15"
        cold_chain:
          type: boolean
          example: false
        beds_available:
          type: number
          example: 4
        staff_present:
          type: number
          example: 3
        staff_sanctioned:
          type: number
          example: 5
...
    OutbreakScenarioRequest:
      type: object
      required: [disease_class, demand_increase_pct, horizon_days]
      properties:
        disease_class:
          $ref: '#/components/schemas/DiseaseClass'
        states:
          type: array
          items:
            type: string
        districts:
          type: array
          items:
            type: string
        demand_increase_pct:
          type: number
          example: 40
        horizon_days:
          type: number
          example: 10
    OutbreakScenarioResult:
      type: object
      required: [scenario_id, is_active, title, description, projected_surge_label, critical_increase_count, affected_phc_ids]
      properties:
        scenario_id:
          type: string
        is_active:
          type: boolean
        title:
          type: string
        description:
          type: string
        projected_surge_label:
          type: string
        critical_increase_count:
          type: number
        affected_phc_ids:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required: [error_code, message, timestamp]
      properties:
        error_code:
          type: string
          example: SCOPE_ACCESS_DENIED
        message:
          type: string
          example: Access denied to requested district resources for role district_officer.
        timestamp:
          type: string
          example: "2026-09-30T10:15:00Z"
        details:
          type: array
          items:
            type: string
  responses:
    400Error:
      description: Invalid parameter or missing required fields.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'